+++ Core/src/main/java/com/gitlab/pauloo27/core/sql/EzDataType.java
@@ -51,6 +51,7 @@ public enum EzDataType {
     // POSTGRESQL
     SERIAL,
     BIGSERIAL,
+    INTERVAL,
     // CUSTOM TYPES
     PRIMARY_KEY;